Replace light monitor by REPL
