+++ lib/dashboard/dashboard.js
@@ -2,7 +2,6 @@ let async = require('async');
 let windowSize = require('window-size');
 
 let Monitor = require('./monitor.js');
-let LightMonitor = require('./light_monitor.js');
 let Console = require('./console.js');
 
 class Dashboard {
@@ -12,7 +11,6 @@ class Dashboard {
     this.plugins = options.plugins;
     this.version = options.version;
     this.env = options.env;
-    this.light = options.light || false;
 
     this.events.on('firstDeploymentDone', this.checkWindowSize.bind(this));
     this.events.on('outputDone', this.checkWindowSize.bind(this));
@@ -39,16 +37,12 @@ class Dashboard {
         callback();
       },
       function startMonitor(callback) {
-        let monitor;
-        if(self.light) {
-          monitor = new LightMonitor({env: self.env, console: console, events: self.events});
-        } else {
-          monitor = new Monitor({env: self.env, console: console, events: self.events});
-          self.events.on('contractsState', monitor.setContracts);
-          self.events.on('status', monitor.setStatus.bind(monitor));
-          self.events.on('servicesState', monitor.availableServices.bind(monitor));
-        }
+        monitor = new Monitor({env: self.env, console: console, events: self.events});
         self.logger.logFunction = monitor.logEntry;
+
+        self.events.on('contractsState', monitor.setContracts);
+        self.events.on('status', monitor.setStatus.bind(monitor));
+        self.events.on('servicesState', monitor.availableServices.bind(monitor));
 
         self.events.setCommandHandler("console:command", monitor.executeCmd.bind(monitor));
