+++ lib/index.js
@@ -240,7 +240,7 @@ class Embark {
 
   console(options) {
     this.context = options.context || [constants.contexts.run, constants.contexts.console];
-    const Dashboard = require('./dashboard/dashboard.js');
+    const REPL = require('./dashboard/repl.js');
     const Engine = require('./core/engine.js');
     const engine = new Engine({
       env: options.env,
@@ -250,58 +250,35 @@ class Embark {
       embarkConfig: options.embarkConfig || 'embark.json',
       logFile: options.logFile,
       logLevel: options.logLevel,
-      context: this.context,
-      useDashboard: true
+      context: this.context
     });
     engine.init();
-    async.parallel([
-      function startDashboard(callback) {
-        let dashboard = new Dashboard({
-          events: engine.events,
-          logger: engine.logger,
-          plugins: engine.plugins,
-          version: engine.version,
-          light: true,
-          env: engine.env
-        });
-        dashboard.start(function () {
-          engine.logger.info(__('dashboard start'));
-          callback();
-        });
-      },
-      function (callback) {
+    async.waterfall([
+      function startServices(callback) {
         let pluginList = engine.plugins.listPlugins();
         if (pluginList.length > 0) {
           engine.logger.info(__("loaded plugins") + ": " + pluginList.join(", "));
         }
 
-        engine.startService("serviceMonitor");
         engine.startService("libraryManager");
         engine.startService("codeRunner");
         engine.startService("web3");
-        engine.startService("pipeline");
-        engine.startService("deployment");
+        engine.startService("deployment", {onlyCompile: true});
         engine.startService("storage");
-        engine.startService("codeGenerator");
-
-        engine.events.on('check:backOnline:Ethereum', function () {
-          engine.logger.info(__('Ethereum node detected') + '..');
-          engine.config.reloadConfig();
-          engine.events.request('deploy:contracts', function (err) {
-            if (err) {
-              return;
-            }
-            engine.logger.info(__('Deployment Done'));
-          });
-        });
-
-        engine.events.on('outputDone', function () {
-          engine.logger.info((__("Looking for documentation? You can find it at") + " ").cyan + "http://embark.status.im/docs/".green.underline + ".".cyan);
-          engine.logger.info(__("Ready").underline);
-          engine.events.emit("status", __("Ready").green);
-        });
 
         callback();
+      },
+      function deploy(callback) {
+        engine.events.request('deploy:contracts', function (err) {
+          engine.logger.info("Finished deploying".underline);
+          callback(err);
+        });
+      },
+      function startREPL(callback) {
+        let repl = new REPL({env: engine.env});
+        repl.start(function () {
+          callback();
+        });
       }
     ], function (err, _result) {
       if (err) {
